[IM] Move common functionality into a trait
