+++ application/controllers/VirtualInterfaceController.php
@@ -33,6 +33,10 @@
  */
 class VirtualInterfaceController extends IXP_Controller_FrontEnd
 {
+    
+    use IXP_Controller_Trait_Interfaces;
+    
+    
     /**
      * This function sets up the frontend controller
      */
@@ -258,7 +262,7 @@ class VirtualInterfaceController extends IXP_Controller_FrontEnd
 	                $vli = new \Entities\VlanInterface();
 	                $form->assignFormToEntity( $vli, $this, false );
 
-                    $vli->setVlan( 
+                    $vli->setVlan(
                         $this->getD2EM()->getRepository( '\\Entities\\Vlan' )->find( $form->getElement( 'vlanid' )->getValue() )
                     );
 
@@ -307,59 +311,5 @@ class VirtualInterfaceController extends IXP_Controller_FrontEnd
         }
     }
 
-    /**
-     * Sets IPv4 or IPv6 from form to given VlanInterface.
-     * 
-     * Function checks if IP address is provided if IP enabled. And then it checks if given IP address exits for current Vlan
-     * if exists then function checks if is not assigned to other virtual interface. If assigned then it add an error and
-     * returns false, else assigns IP to Vlan interface. If not IP not existent function creates new one and assign it to Vlan
-     * interface.
-     *
-     * @param IXP_Form_Interface_AddWizard $form Form to get IP address data
-     * @param \Entities\VirtualInterface   $vi   Virtual interface to check if IP is not assign to different
-     * @param \Entities\VlanInterface      $vli  Vlan interface to assign IP to
-     * @param bool                         $ipv6 Flat to define if IP address is IPv4 or IPv6
-     * @return bool
-     */
-    private function setIp( $form, $vi, $vli, $ipv6 = false )
-    {
-        $iptype = $ipv6 ? "ipv6" : "ipv4";
-        $ipVer = $ipv6 ? "IPv6" : "IPv4";
-        $setter = "set{$ipVer}Address";
-        $entity = "\\Entities\\{$ipVer}Address";
-        $vlan = $vli->getVlan();
-
-        if( $form->getValue( $iptype . "enabled" ) )
-        {
-            if( !$form->getElement( $iptype . 'addressid' )->getValue() )
-            {
-                $form->getElement( $iptype . 'addressid' )->setErrors( ["Please select or enter a {$ipVer} address."] );
-                return false;
-            }
-
-            $ip = $this->getD2EM()->getRepository( $entity )->findOneBy( [ "Vlan" => $vlan->getId(), 'address' => $form->getElement( $iptype . 'addressid' )->getValue() ] );
-            if( !$ip )
-            {
-                $ip = new $entity();
-                $this->getD2EM()->persist( $ip );
-                $ip->setVlan( $vlan );
-                $ip->setAddress( $form->getElement( $iptype . 'addressid' )->getValue() );
-            }
-            else
-            {
-                //If VlanInterface is assign to ip and Vlan interface is different then it returns false
-                if( $ip->getVlanInterface() && $ip->getVlanInterface() != $vli )
-                {
-                    $address = $form->getElement( $iptype . 'addressid' )->getValue();
-                    $form->getElement( $iptype . 'addressid' )->setErrors( ["{$ipVer} address '{$address}' is already in use." ] );
-                    return false;
-                }
-
-            }
-            $vli->$setter( $ip );
-        }
-        return true;
-    }
-        
 }
